Refactored code for further SMT impl.
